+++ CHANGELOG.rst
@@ -44,6 +44,8 @@ Fixed
 
 Changed
 ^^^^^^^
+* `@senwu`_: Change default grad clip to None.
+* `@senwu`_: Update seed and grad_clip to nullable.
 * `@senwu`_: Change default class index to 0-index.
 * `@senwu`_: Change default ignore_index to None.
 * `@senwu`_: Change the default counter unit to epoch.